Incorporating the lastest changes from my private git projects. Major change:   - Implemented residual normalization (in addition to EMSC and PEMSC)
I did not thoroughly test the new code, but I'm quite optimistic that it
is still "interpretable" (i.e., does not raise runtime exception---for
example, due to incorrect method signatures).

Besides, I change the evaluation scripts to evaluate residual
normalization. The results suggest that is can yield good results
(e.g., if there is no residual probability, yet this was expected) but
is less stable in general. For example, there are more models for which
I cannot find good weights (on average) within a reasonable number of repetitions.
Moreover, the quality of the resulting model exhibits stronger variance.